requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
psycopg==3.2.9
django-cors-headers==4.7.0
drf-spectacular==0.28.0
django-environ==0.12.0
openai>=1.30.0,<2.0.0
scikit-learn==1.5.2
pgvector==0.2.5
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.4
faiss-cpu==1.13.0
chromadb==0.5.3
pypdf==5.1.0
geopy==2.4.1
requests-html==0.10.0
beautifulsoup4==4.12.3
pandas==2.2.3
argon2-cffi==25.1.0
redis==6.2.0
django-redis==5.4.0
celery==5.5.3
django-celery-results==2.6.0
channels==4.2.2
channels-redis==4.2.1
uvicorn[standard]==0.35.0
requests==2.32.4
aiohttp==3.12.13
orjson==3.10.18
msgpack==1.1.1
python-dateutil>=2.8.2
phonenumbers==9.0.8
pillow==11.3.0
django-jazzmin==3.0.1
gunicorn==23.0.0